MDEV-4786 - merge 10.0-monty - 10.0
Fixed failure of one of the test case from innodb_ext_key.test.
The fact is that the innodb code may return statistical data
on record per key values that is far off the real numbers. 
This is exactly what happened in mariadb-5.5 with this test case.

Added an ANALYZE command in this test case to avoid this problem.
The same change will be done in mariadb-5.5