MDEV-5042 - Server crashes when accessing incorrect MERGE table from trigger
When we open merge children open error is normally handled early during
open phase. But there are two exceptions when error is handled later during
attach phase: CHECK/REPAIR TABLE and tables added by the pre-locking code.

The latter case wasn't considered by assertion in the merge code. This
assertion is corrected. Note that in MySQL-5.6 this assertion is removed.